Implement 04-collisions, double hashing
Added the function for double hashing the key. This is done to avoid
collisions. Collisions occur because there is only a finite amount of
outputs, remember 0 to m. Different inputs map to the same array index.

Needed to add HT_PRIME_* constants which are not outlined in the
tutorial but were found by looking at the original authors
implementation.