Implement 05-methods
Implemented search, delete, and insert functions for the hash table.

Needed to change the strdup to malloc and strcpy. Apparently strdup is
not part of c99 and was not actually working. Not 100% on all the
reasons (stacked overflow).

Changed the comments to /* */ during attempts to resolve the strdup
(switched to c89 compiler).