Implementation of 06-resizing
For the resizing implementation, compared to the tutorial on github had
to change a few things.
First, xmalloc/xcalloc is not defined. It is not part of the std library
in c, and just kills the program if malloc/calloc fails. For this
tutorial not important so used malloc/calloc instead.
Second, HT_INITIAL_BASE_SIZE is not shown to be defined so added that.
Third, the member variable size_index is added to struct ht_hash_table.